test(integration): consolidate uptodate and outOfDate into one suite
Both tests are the same 'Uptodate' suite over the project-uptodate workspace with
identical setup; merge outOfDate's case into uptodate.test.ts as a second it() so
they share one Electron instance and one authentication/setup cycle. Neither uses
sinon, so there is no cross-suite state to leak.

Integration test files: 24 -> 23.
