+++ test/integration/uptodate/uptodate.test.ts
@@ -20,6 +20,23 @@ describe('Uptodate', () => {
   beforeEach(waitForIndexer);
   afterEach(initializeWorkspace);
 
+  it('detects when the AppMap is out of date', async () => {
+    const uptodateService = await waitForDependsUpdate();
+    const serviceInstances = (await waitForExtension()).workspaceServices.getServiceInstances(
+      uptodateService
+    );
+    assert.strictEqual(serviceInstances.length, 1);
+
+    await repeatUntil(
+      () => touch(UserFile),
+      `${UserPageAppMapFile} is still considered up to date`,
+      () => !uptodateService.isUpToDate(vscode.Uri.file(UserPageAppMapFile))
+    );
+
+    const outOfDateTestLocations = await uptodateService.outOfDateTestLocations();
+    assert.deepStrictEqual(outOfDateTestLocations, ['spec/requests/user_spec.rb:10']);
+  });
+
   it('detects when the AppMap is up to date again', async () => {
     const uptodateService = await waitForDependsUpdate();
     const serviceInstances = (await waitForExtension()).workspaceServices.getServiceInstances(
